continue adapt slide with vars
