+++ src/components/App/Exam/Exam.js
@@ -1,7 +1,7 @@
 import './exam.scss'
 
 import React, { useState } from 'react'
-import { useDispatch } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
 
 import {
   Paper,
@@ -15,7 +15,7 @@ import TextField from '@material-ui/core/TextField'
 import ArrowBackIcon from '@material-ui/icons/ArrowBack'
 import ArrowForwardIcon from '@material-ui/icons/ArrowForward'
 import InputLabel from '@material-ui/core/InputLabel'
-
+import MenuItem from '@material-ui/core/MenuItem'
 import imgDesktop from '../../../images/illus.png'
 import imgMobile from '../../../images/mobile-bg.svg'
 import oStyle from '../../ResponsiveDesign/AuthStyle'
@@ -30,13 +30,14 @@ const useStyles = makeStyles((theme) => (oStyle(theme, imgDesktop, imgMobile)))
 const Exam = (props) => {
   const classes = useStyles()
   const dispatch = useDispatch()
+  const symptomes = useSelector((state) => state.home.symptomes)
+
   const initVals = {
     exam_pics: [],
     extra_exam_desc: [],
     errIntra_extra_oral_desc: false,
-    //errProblem_health: false,
-    //errIn_treatment: false,
     errExtra_exam_name: false,
+    errSymptomes: false
   }
   const [errors, setErrors] = useState(initVals)
 
@@ -44,13 +45,10 @@ const Exam = (props) => {
     event.preventDefault()
     let isValid = true
     if (props.values.intra_extra_oral_desc === '') { setErrors({ ...errors, errIntra_extra_oral_desc: true }); isValid = false }
-    if (props.values.problem_health === '') { setErrors({ ...errors, errProblem_health: true }); isValid = false }
-    if (props.values.in_treatment === '') { setErrors({ ...errors, errIn_treatment: true }); isValid = false }
-    if (props.values.extra_exam.extra_exam_name === '') { setErrors({ ...errors, errExtra_exam_name: true }); isValid = false }
-    if (props.values.extra_exam.extra_exam_desc === '') { setErrors({ ...errors, errExtra_exam_desc: true }); isValid = false }
-
-    
-
+    if (props.values.extra_exam.extra_exam_name < 1) { setErrors({ ...errors, errExtra_exam_name: true }); isValid = false }
+    if (props.values.extra_exam.extra_exam_desc < 1) { setErrors({ ...errors, errExtra_exam_desc: true }); isValid = false }
+    if (props.values.symptomes.length < 1) { setErrors({ ...errors, errSymptomes: true }); isValid = false }
+   
     if (isValid) { dispatch({ type: UPDATE_LEVEL, level: 'diagnostic' }) }
   }
 
@@ -115,6 +113,34 @@ const Exam = (props) => {
                 error={errors.errIntra_extra_oral_desc}
               />
 
+              <br /> <br />
+
+              <Typography component='h1' variant='h5'>
+                Symptomes
+              </Typography>
+              <TextField
+                className='textField'
+                id='Symptomes'
+                label='Symptomes'
+                select
+                onChange={props.onChange('symptomes')}
+                variant='outlined'
+                fullWidth
+                SelectProps={{
+                  multiple: true,
+                  value: props.values.symptomes
+                }}
+                error={errors.errSymptomes}
+              >
+                {symptomes && symptomes.map((value) => (
+                  <MenuItem key={value['@id']} value={value['@id']}>
+                    {value.name}
+                  </MenuItem>
+                ))}
+              </TextField>
+
+              <br /> <br />
+
               <Typography component='h1' variant='h5'>
                 Examen Complementaire
               </Typography>
@@ -148,7 +174,7 @@ const Exam = (props) => {
                 name='extra_exam_name'
                 type='textarea'
                 id='extra_exam_name'
-                value={props.values.extra_exam.extra_exam_name}
+                value={props.values.extra_exam.extra_exam_name[0]}
                 autoComplete='current-extra_exam_name'
                 onChange={props.onChange('extra_exam_name')}
                 error={errors.errExtra_exam_name}
@@ -167,7 +193,7 @@ const Exam = (props) => {
                 name='extra_exam_desc'
                 type='textarea'
                 id='extra_exam_desc'
-                value={props.values.extra_exam.extra_exam_desc}
+                value={props.values.extra_exam.extra_exam_desc[0]}
                 autoComplete='current-extra_exam_desc'
                 onChange={props.onChange('extra_exam_desc')}
                 error={errors.errExtra_exam_desc}
